Split freight_traffic.rs (580 lines) into freight_traffic/ subdirectory
Split into focused modules:
- constants.rs: all tuning constants
- types.rs: FreightTruck, FreightTrafficState, save data, Saveable impl
- systems.rs: demand computation, trip generation, truck movement, satisfaction
- plugin.rs: FreightTrafficPlugin registration
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports
